Pull most of the trajectory stuff into a reusable logic component that other commands could use (for example, if they wanted to follow successive routes)
